Added Token_stream.h and Token_stream.cpp, updated #includes
