+++ Calculator/Calculator.vcxproj
@@ -149,12 +149,14 @@
   <ItemGroup>
     <ClCompile Include="Calculator.cpp" />
     <ClCompile Include="Functions.cpp" />
+    <ClCompile Include="Token_stream.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Errors.h" />
     <ClInclude Include="Functions.h" />
     <ClInclude Include="Gconst.h" />
     <ClInclude Include="Token.h" />
+    <ClInclude Include="Token_stream.h" />
     <ClInclude Include="Variable.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
