+++ Calculator/Token.h
@@ -2,6 +2,8 @@
 
 #include <string>
 
+using std::string;
+
 //------------------------------------------------------------------------------
 // Defines 3 Tokens of different types: a character token "Token(char ch)" with no value,
 // a number token "Token(char ch, double val)" with an associated numeric value,
